// src/generated/models/system-user-model.ts
import type { User } from './user-model';

export const SystemUserRoleKeyToLabel = {
  'Admin': 'Admin',
  'ProjectManager': 'Project Manager',
  'Approver': 'Approver',
  'Employee': 'Employee'
} as const;
export type SystemUserRoleKey = keyof typeof SystemUserRoleKeyToLabel;

export interface SystemUser {
  /**
   * @displayName System User
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Full Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  fullName: string;
  /**
   * @displayName Email
   */
  email?: string;
  /**
   * @displayName Is Active
   */
  isActive?: boolean;
  /**
   * @displayName Role
   * @validationRule Required for create/update operations
   */
  roleKey: SystemUserRoleKey;
  /**
   * @displayName User
   * @validationRule Required for create/update operations
   */
  user: Pick<User, 'id' | 'fullName'>;
}

export const _SystemUser = 'SystemUser' as const;